debug: Refactor to separate constants and output functions
