+++ talk4_debug/demoLocalDebug/bad/utils/constants.py
@@ -1,0 +1,2 @@
+KB = 1.024
+MB == 1024 * KB
