+++ talk4_debug/demoLocalDebug/good/classify.py
@@ -1,8 +1,7 @@
 import os
-from infolib import FileInfo
+from utils import FileInfo, exportCSV
+from utils.constants import MB
 
-KB = 1024
-MB = 1024 * KB
 
 DIRECTORY = "./images"
 SIZE_LIMIT = 0.1 * MB
@@ -21,60 +20,27 @@ def getFiles(
 		files.append(file)
 	return files
 
-def getTotalSize(
-		files: list
-	) -> int:
-	"""Obtain total size of files in the list."""
-
-	totalSize = 0
-	for index, file in enumerate(files):
-		totalSize += file.size
-	return totalSize
-
-def filterBySize(
+def classifyBySize(
 		files: list,
 		maxSize: int = 2**40  # 1 TB
 	) -> tuple:
-	"""Filter a list of files based on their size.
+	"""Classify a list of files based on their size.
 	
-	Returns a tuple of 2 lists with included and excluded files.
+	Returns a tuple of 2 lists with small and big files.
 	"""
 	
-	included = []
-	excluded = []
+	smallFiles = []
+	bigFiles = []
 	for file in files:
 		if file.size <= maxSize:
-			included.append(file)
+			smallFiles.append(file)
 		else:
-			excluded.append(file)
-	return (included, excluded)
-
-def exportCSV(
-		files: list,
-		csvFile: str = "output.csv"
-	) -> None:
-	"""Export CSV from list of files."""
-	
-	csvHeaders = "basename\tsize\tdatetime\n"
-	csvData = [
-			f"{file.basename}\t{file.size}\t{file.format_datetime()}\n"
-			for file in files
-	]
-
-	# write data to output file
-	print(f"Writing '{csvFile}'...")
-	with open("csv" + os.sep + csvFile, "w") as output:
-		output.write(csvHeaders)
-		output.writelines(csvData)
-	
-	# print summary
-	print(f"{len(files)} files.")
-	totalMB = getTotalSize(files) / MB
-	print(f"{totalMB:0.2} MB in files")
+			bigFiles.append(file)
+	return (smallFiles, bigFiles)
 
 if __name__ == "__main__":
 	images = getFiles(DIRECTORY)
-	small, large = filterBySize(images, maxSize=SIZE_LIMIT)
+	small, large = classifyBySize(images, maxSize=SIZE_LIMIT)
 	exportCSV(small, csvFile="small.csv")
 	exportCSV(large, csvFile="large.csv")
