+++ talk4_debug/demoLocalDebug/good/utils/constants.py
@@ -1,0 +1,2 @@
+KB = 1024
+MB = 1024 * KB
